chore(main.yml): update GitHub Actions workflow to use the latest version of actions/checkout and gluonhq/setup-graalvm
The previous version of the workflow was using actions/checkout@v2 and actions/setup-java@v2. This commit updates the workflow to use actions/checkout@v3 and gluonhq/setup-graalvm@master.

The changes also include setting up Gluon's GraalVM with the version 22.1.0.1-Final and granting execute permission for gradlew.

These updates were made to ensure that the workflow is using the latest versions of the actions and to take advantage of the latest features and improvements provided by these actions.